@@ -1,31 +1,16 @@
 package com.hendrix.test;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-
-import java.util.concurrent.ThreadPoolExecutor;
-
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.Test;
 
 /**
  * Created by tshalev on 09/02/2017.
  */
 public class GraphTestsTest {
-    @Before
-    public void setUp() throws Exception {
-
-    }
-
-    @After
-    public void tearDown() throws Exception {
-
-    }
 
     @Test
     public void testHello() {
-
-        assert (true);
+        assertTrue(true);
     }
 
 }
